refactored the action-controllers and spilted the code into smaller chunks for better structure
