+++ src/controllers/user-actions/inviteController.js
@@ -1,8 +1,8 @@
-const Invite = require("../models/Invite");
-const User = require("../models/User");
-const Workspace = require("../models/Workspace");
-const Channels = require("../models/Channel");
-const asyncErrorHandler = require("../utils/AsyncErrorHandler");
+const Invite = require("../../models/Invite");
+const User = require("../../models/User");
+const Workspace = require("../../models/Workspace");
+const Channels = require("../../models/Channel");
+const asyncErrorHandler = require("../../utils/AsyncErrorHandler");
 
 // Endpoint
 exports.generateInviteLink = asyncErrorHandler(async (req, res, next) => {
