+++ src/controllers/user-auth/otpController.js
@@ -1,5 +1,5 @@
-const OtpModel = require("../models/Otp")
-const emailSender = require("../utils/emailSender")
+const OtpModel = require("../../models/Otp")
+const emailSender = require("../../utils/emailSender")
 
 function generateOtp () {
     return Math.floor(100000 + Math.random() * 900000).toString();
