+++ src/router/user-actions/feedBackRouter.js
@@ -1,0 +1,12 @@
+const express = require("express");
+const feedBackController = require("../../controllers/user-actions/feedBackController");
+const authController = require("../../controllers/user-auth/authController");
+
+const router = express.Router();
+
+// FEEDBACK
+router
+  .route("/api/v1/feedback")
+  .post(authController.protect, feedBackController.addFeedback);
+
+module.exports = router;
