+++ src/router/user-auth/authRoute.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const authController = require("../controllers/authController");
+const authController = require("../../controllers/user-auth/authController");
 
 const router = express.Router();
